extract getAddressableAreaFromConfig helper
